acpi_einj: Support for ACPI error injection
This driver parses the ACPI EINJ table and builds a list of
instructions associated with known actions.  It then exports ioctls to
execute specific sequences of actions to permit injecting system
errors.  This can be used to test error reporting facilities for
events such as ECC errors.

Currently this driver supports a single ioctl to query the set of
available error types.

Sponsored by:	Netflix