iw_cxgbe/libcxgb4: Support 64B CQEs
The larger CQE exports the RSS header to userspace.  Note that this
changes the ABI between the driver and library, so the ABI version is
bumped.  A new user->kernel create_cq request message is defined to
disallow a CQE size mismatch between iw_cxgbe and libcxgb4.

Sponsored by:	Chelsio Communications
Co-authored-by:	Steve Wise <swise@opengridcomputing.com>