bus: Stop invoking BUS_PROBE_NOMATCH for kldunload
Originally, BUS_PROBE_NOMATCH was only invoked when device_probe
failed to locate a driver for a device.  It's original purpose was to
announce devices without a driver (e.g. enumerating such devices found
during the boot-time device probe).

Over time, some bus drivers added additional logic to this hook.
Notably, the PCI bus gained a knob to optionally power down devices
without a driver.

Commit e7d83347c01280e7021144722d7c30704de2a276 in June of 2010 added
an invocation of BUS_PROBE_NOMATCH in devclass_delete_driver so that
devices could be powered down if their driver was unloaded.  It also
reset the DF_DONENOMATCH flag after a successfull attach, though this
has some curious semantics.  Namely, if a device does an attach/detach
cycle where the detach was triggered by unloading the driver, the flag
was ignored anyway as the BUS_PROBE_NOMATCH call was unconditional in
devclass_delete_driver.  However, if the device was detached for some
other reason, and later failed to re-probe, that failing probe would
invoke BUS_PROBE_NOMATCH, but at the time of re-probing, not when the
device was detached.

Three years later in commit e35ce1f271aba87a543fc94219e11a594aba99d3,
I added a BUS_CHILD_DETACHED implementation to the PCI bus driver that
powers down devices anytime a PCI device is detached whether by
kldunload or other means.  This renders the BUS_PROBE_NOMATCH
invocation in the kldunload case redundant as the only effect it has
now is to add additional clutter to dmesg announcing that the device
that has just been detached does not have a driver.

This commit reverts the first part of
e7d83347c01280e7021144722d7c30704de2a276, but not the second part.
Note that due to the second part, devmatch added yet another flag
(DF_ATTACHED_ONCE) to avoid a doom loop of kldunload triggering
nomatch events that would try to re-load the module just unloaded.  If
we reverted the second part of this commit, we would avoid sending
nomatch events due to kldunload entirely (or when a device is detached
for another reason such as a devctl(8) request).