freebsd32_setcred: Copy all of the setcred fields
The previous memcpy did not actually work because the size expression
was the difference of two pointers to integers, so it only copied
1/4th of the relevant bytes and used stack garbage for the remaining
bytes.  This could be fixed by using the __rangeof() macro, but using
CP() directly is simpler and more typical of other freebsd32 system
calls.  Modern compilers are smart enough to coalesce multiple member
assignments into a bulk copy.

Obtained from:	CheriBSD