netlink: Optimize growing the snl_writer message buffer
Optimize the case that the current write buffer grown by
snl_realloc_msg_buffer is allocated at the end of the associated
snl_state's current linear buffer if the linear buffer has room for
the new size.  In that case, simply reuse the existing space in
the linear buffer by bumping the offset of the linear buffer for
the updated size.

Sponsored by:	AFRL, DARPA